26-04-2026 update version

+++ src/Util.java
@@ -2,38 +2,46 @@ import java.io.File;
 
 public class Util {
 
-    public static void createSubfolder( String folderName){
-        String directory="C:\\Users\\user\\Documents\\io\\"+folderName;
-        File subFolder = new File(directory);
-        if(!subFolder.exists()){
+    public static void createSubfolder(File mainFolder, String folderName) {
+        File subFolder = new File(mainFolder, folderName);
+        if (!subFolder.exists()) {
             subFolder.mkdir();
         }
     }
 
-    public static void createNestedSubfolder( String folderNames){
-        String directory="C:\\Users\\user\\Documents\\io\\"+folderNames;
-        File subFolders = new File(directory);
-        if(!subFolders.exists()){
+    public static void createNestedSubfolder(File mainFolder, String folderNames) {
+        File subFolders = new File(mainFolder, folderNames);
+        if (!subFolders.exists()) {
             subFolders.mkdirs();
         }
     }
 
-    public static void folderHierarchy() {
-        File files = new File("C:\\Users\\user\\Documents\\io");
-        File[] file = files.listFiles();
-        for (File f : file) {
-            if (f.isDirectory()) {
-                System.out.println(f.getName());
+    public static void folderHierarchy(File mainFolder) {
+        File[] file = mainFolder.listFiles();
+        if (file != null) {
+            for (File f : file) {
+                if (f.isDirectory()) {
+                    System.out.println("Qovluq: " + f.getName());
+                }
             }
+        } else {
+            System.out.println("Göstərilən qovluq tapılmadı");
         }
     }
 
-    public static void deleteFolder(String folderName){
-        String directory="C:\\Users\\user\\Documents\\io\\"+folderName;
-        File subFolders = new File(directory);
-        subFolders.delete();
+    public static void deleteFolder(File mainFolder, String folderName) {
+        File subFolders = new File(mainFolder, folderName);
+        if (subFolders.exists()) {
+            boolean flag = subFolders.delete();
+            if (flag) {
+                System.out.println("Qovluq silindi");
+            } else {
+                System.out.println("Qovluq silinmədi. Daxilində fayl ola bilər.");
+            }
+        }else{
+            System.out.println("Qovluq mövcud deyil.");
+        }
     }
-
-
-
 }
+
+
